+++ tlm_adjoint/tlm_adjoint.py
@@ -46,24 +46,7 @@ __all__ = \
         "CheckpointStorage",
         "Control",
         "EquationManager",
-        "ManagerException",
-        "add_tlm",
-        "annotation_enabled",
-        "compute_gradient",
-        "configure_checkpointing",
-        "manager_info",
-        "new_block",
-        "reset",
-        "reset_adjoint",
-        "reset_manager",
-        "start_annotating",
-        "start_manager",
-        "start_tlm",
-        "stop_annotating",
-        "stop_manager",
-        "stop_tlm",
-        "tlm",
-        "tlm_enabled"
+        "ManagerException"
     ]
 
 
@@ -1703,113 +1686,3 @@ class EquationManager:
 
 
 set_manager(EquationManager())
-
-
-def configure_checkpointing(cp_method, cp_parameters={}, manager=None):
-    if manager is None:
-        manager = _manager()
-    manager.configure_checkpointing(cp_method, cp_parameters=cp_parameters)
-
-
-def manager_info(info=print, manager=None):
-    if manager is None:
-        manager = _manager()
-    manager.info(info=info)
-
-
-def reset_manager(cp_method=None, cp_parameters=None, manager=None):
-    if manager is None:
-        manager = _manager()
-    manager.reset(cp_method=cp_method, cp_parameters=cp_parameters)
-
-
-def reset(cp_method=None, cp_parameters=None, manager=None):
-    warnings.warn("reset is deprecated -- use reset_manager instead",
-                  DeprecationWarning, stacklevel=2)
-    if manager is None:
-        manager = _manager()
-    manager.reset(cp_method=cp_method, cp_parameters=cp_parameters)
-
-
-def annotation_enabled(manager=None):
-    if manager is None:
-        manager = _manager()
-    return manager.annotation_enabled()
-
-
-def start_manager(annotation=True, tlm=True, manager=None):
-    if manager is None:
-        manager = _manager()
-    manager.start(annotation=annotation, tlm=tlm)
-
-
-def start_annotating(manager=None):
-    if manager is None:
-        manager = _manager()
-    manager.start(annotation=True, tlm=False)
-
-
-def start_tlm(manager=None):
-    if manager is None:
-        manager = _manager()
-    manager.start(annotation=False, tlm=True)
-
-
-def stop_manager(annotation=True, tlm=True, manager=None):
-    if manager is None:
-        manager = _manager()
-    manager.stop(annotation=annotation, tlm=tlm)
-
-
-def stop_annotating(manager=None):
-    if manager is None:
-        manager = _manager()
-    manager.stop(annotation=True, tlm=False)
-
-
-def stop_tlm(manager=None):
-    if manager is None:
-        manager = _manager()
-    manager.stop(annotation=False, tlm=True)
-
-
-def add_tlm(M, dM, max_depth=1, manager=None):
-    if manager is None:
-        manager = _manager()
-    manager.add_tlm(M, dM, max_depth=max_depth)
-
-
-def tlm_enabled(manager=None):
-    if manager is None:
-        manager = _manager()
-    return manager.tlm_enabled()
-
-
-def tlm(M, dM, x, max_depth=1, manager=None):
-    if manager is None:
-        manager = _manager()
-    return manager.tlm(M, dM, x, max_depth=max_depth)
-
-
-def reset_adjoint(manager=None):
-    warnings.warn("reset_adjoint is deprecated",
-                  DeprecationWarning, stacklevel=2)
-    if manager is None:
-        manager = _manager()
-    manager.reset_adjoint(_warning=False)
-
-
-def compute_gradient(Js, M, callback=None, prune_forward=True,
-                     prune_adjoint=True, adj_ics=None, manager=None):
-    if manager is None:
-        manager = _manager()
-    return manager.compute_gradient(Js, M, callback=callback,
-                                    prune_forward=prune_forward,
-                                    prune_adjoint=prune_adjoint,
-                                    adj_ics=adj_ics)
-
-
-def new_block(manager=None):
-    if manager is None:
-        manager = _manager()
-    manager.new_block()
